JamJelly2/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autocommand==2.2.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
df2gspread==1.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
google-api-python-client==1.6.7
google-auth==1.28.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mintapi==1.44
mock==4.0.3
numpy==1.20.1
oathtool==2.3.0
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.2.2
path==15.1.2
pipenv==2020.11.15
plotly==4.14.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
selenium==3.141.0
selenium-requests==1.3
six==1.15.0
spotipy==2.16.1
sqlparse==0.4.1
tldextract==3.1.0
twilio==6.57.0
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==1.0.1
whitenoise==5.2.0
xmltodict==0.12.0
